lib/ukbitops: Move bitmap funcs out of bitops.h
This change moves all bitmap-related functionality out of the main
bitops.h into its own header.

Checkpatch-Ignore: PARENTHESIS_ALIGNMENT
Checkpatch-Ignore: VOLATILE
Checkpatch-Ignore: SPACING
Checkpatch-Ignore: UNNECESSARY_PARENTHESES
Checkpatch-Ignore: MACRO_ARG_REUSE
